+++ src/pages/Tools/OneKV/components/oneKVStatus.tsx
@@ -1,19 +1,17 @@
-import { useCallback, useEffect, useMemo, useState } from 'react';
+import { useEffect, useState } from 'react';
 import styled from 'styled-components';
 import moment from 'moment';
-import { ReactComponent as DashboardIcon } from '../../../../assets/images/dashboard.svg';
+
 import { ReactComponent as MonitorIcon } from '../../../../assets/images/monitor.svg';
 import { ReactComponent as Search } from '../../../../assets/images/search.svg';
-import { ReactComponent as ActiveIcon } from '../../../../assets/images/active.svg';
-import { ReactComponent as InactiveIcon } from '../../../../assets/images/inactive.svg';
+
 import CardHeader from '../../../../components/Card/CardHeader';
 import IconInput from '../../../../components/Input/IconInput';
-import Table from '../../../../components/Table';
 import { useAppSelector } from '../../../../hooks';
-import { formatBalance } from '@polkadot/util';
+
 import { apiGetAllOneKVValidator, IOneKVValidator, IOneKVValidators } from '../../../../apis/OneKV/validator';
-import { useHistory } from 'react-router-dom';
 import { apiGetOneKVNominators, IOneKVNominators } from '../../../../apis/OneKV/nominator';
+import ValidatorTable from './oneKVValidTable';
 
 const OneKVHeader = () => {
   return (
@@ -26,138 +24,6 @@ const OneKVHeader = () => {
         </HeaderTitle>
       </HeaderLeft>
     </HeaderLayout>
-  );
-};
-
-const ValidatorTable = ({filter, chain, validators}) => {
-  const history = useHistory();
-  const onClickDashboard = useCallback((id: string) => {
-    history.push(`/tools/validator/${id}/${chain}`);
-  }, [chain, history]);
-  const _formatBalance = useCallback((value: any) => {
-    if (chain === 'KSM') {
-      return (<span>{formatBalance(value, {
-        decimals: 12,
-        withUnit: 'KSM'
-      })}</span>);
-    } else if (chain === 'DOT') {
-      return (<span>{formatBalance(value, {
-        decimals: 10,
-        withUnit: 'DOT'
-      })}</span>);
-    }
-  }, [chain]);
-  const columns = useMemo(() => {
-    return [
-      {
-        Header: 'Dashboard',
-        accessor: 'dashboard',
-        maxWidth: 48,
-        disableSortBy: true,
-        Cell: ( {row} ) => {
-          return(<span><DashboardIcon 
-            onClick={() => {onClickDashboard(row.original.stash)}}/></span>);
-        },
-      },
-      {
-        Header: 'Name',
-        accessor: 'name',
-        maxWidth: 180,
-        Cell: ({ value }) => <span>{value}</span>,
-      },
-      {
-        Header: 'Commission',
-        accessor: 'stakingInfo.validatorPrefs.commission',
-        maxWidth: 48,
-        Cell: ({ value }) => <span>{value / 10000000}%</span>,
-      },
-      {
-        Header: 'Active',
-        accessor: 'activeNominators',
-        maxWidth: 60,
-        Cell: ({ value }) => {
-          if (value > 0) {
-            return (<span><ActiveIcon /></span>);
-          } else {
-            return (<span><InactiveIcon /></span>);
-          }
-        },
-      },
-      {
-        Header: '1KV nominated',
-        accessor: 'elected',
-        maxWidth: 100,
-        Cell: ({ value, row }) => {
-          if (value  === true) {
-            return (<span><ActiveIcon /></span>);
-          } else {
-            console.log(row);
-            return (
-            <OneKVNominated>
-              <InactiveIcon />
-              <LastNominationDate>({moment(row.original.nominatedAt).format('MM/DD')})</LastNominationDate>
-            </OneKVNominated>);
-          }
-        },
-      },
-      {
-        Header: 'Nomination Order',
-        accessor: 'nominationOrder',
-        maxWidth: 60,
-        Cell: ({ value }) => {
-          return (<span>{value}</span>);
-        },
-      },
-      {
-        Header: 'Self Stake',
-        accessor: 'selfStake',
-        maxWidth: 150,
-        Cell: ({ value }) => {
-          return (<span>{_formatBalance(value)}</span>);
-        }
-      },
-      {
-        Header: 'Rank',
-        accessor: 'rank',
-        maxWidth: 60,
-        Cell: ({ value }) => {
-          return (<span>{value}</span>);
-        },
-      },
-      {
-        Header: 'Inclusion',
-        accessor: 'inclusion',
-        maxWidth: 60,
-        Cell: ({ value }) => {
-          return (<span>{(value * 100).toFixed(2)}%</span>);
-        },
-      },
-    ]
-  }, [_formatBalance, onClickDashboard]);
-  const [displayValidators, setDisplayValidators] = useState<IOneKVValidator[]>([]);
-  useEffect(() => {
-    setDisplayValidators(validators);
-  }, [chain, validators]);
-  useMemo(() => {
-    if (filter.stashId.length > 0) {
-      const displayValidators: IOneKVValidator[] = [];
-      validators.forEach((v) => {
-        if (v.stash.toLowerCase().includes(filter.stashId.toLowerCase())) {
-          displayValidators.push(v);
-        } else if(v.name.toLowerCase().includes(filter.stashId.toLowerCase())) {
-          displayValidators.push(v);
-        }
-      });
-      setDisplayValidators(displayValidators);
-    } else {
-      setDisplayValidators(validators);
-    }
-  }, [filter.stashId, validators]);
-  return (
-    <Table
-      columns={columns}
-      data={displayValidators}
-    />
   );
 };
 
@@ -326,19 +192,6 @@ const OptionBar = styled.div`
   padding: 12px 0px 0px 13.8px;
   border-radius: 6px;
   background-color: #2f3842;
-`;
-
-const OneKVNominated = styled.div`
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-`;
-
-const LastNominationDate = styled.div`
-  margin: 0 0 0 4px;
-  justify-content: center;
-  align-items: center;
 `;
 
 const HeaderItem = styled.div`
